add medication use case first step ( save mediaction image )

+++ usecase/entity/medication.go
@@ -2,7 +2,7 @@ package entity
 
 type Medication struct {
 	Name   string
-	Weight int64
+	Weight float64
 	Code   string
 	Image  string
 }
